refactor: Update Crucial.StorageExecutive.ps1 to use wingetcreate for updating package
This commit updates the Crucial.StorageExecutive.ps1 script to use the wingetcreate tool for updating the package. It replaces the previous method of using Komac. The script now downloads wingetcreate.exe and uses it to update the package with the latest version and URLs. This change improves the update process for Crucial.StorageExecutive package.